Unixify unsquashfs, half-way to become an api and an r2 plug

+++ libr/fs/squashfs/unsquash-4.c
@@ -102,7 +102,7 @@ struct inode *read_inode_4(unsigned int start_block, unsigned int offset)
 
 	if (bytes == -1) {
 		fprintf (stderr, "read_inode: inode table block %lld not found\n", start);
-		fprintf (stderr, "::: bytes -1 offset = %lld\n", offset);
+		fprintf (stderr, "::: bytes -1 offset = %lld\n", (uint64_t)offset);
 		return NULL;
 	}
 
@@ -240,7 +240,7 @@ struct inode *read_inode_4(unsigned int start_block, unsigned int offset)
 		}
 		default:
 			fprintf(stderr, "Unknown inode type %d in read_inode!\n", header.base.inode_type);
-fprintf (stderr, "bytes -1\n");
+			fprintf (stderr, "bytes -1\n");
 			return NULL;
 	}
 	return &i;
